backend/run_simple.py
#!/usr/bin/env python3
"""
Startup script for the Internship Allocation Engine backend (simple version)
"""
import uvicorn
import os
from dotenv import load_dotenv

load_dotenv()

if __name__ == "__main__":
    uvicorn.run(
        "app.main_simple:app",
        host="0.0.0.0",
        port=8000,
        reload=True,
        log_level="info"
    )
